added static style in some blocks
