+++ blocks-config/uagb-controls/AddStaticStyles.js
@@ -1,0 +1,13 @@
+import { useEffect } from '@wordpress/element';
+import { useDeviceType } from '@Controls/getPreviewType';
+const AddStaticStyles = ( ChildComponent )=> {
+	return ( props ) => {
+		const deviceType = useDeviceType();
+		useEffect( () => {
+			addBlockEditorDynamicStyles();
+		}, [ deviceType ] );
+	
+		return <ChildComponent { ...props }/>
+	}
+}
+export default AddStaticStyles;
