+++ src/blocks/timeline/post-timeline/edit.js
@@ -4,7 +4,6 @@
 
 import { useEffect, useState, useMemo } from '@wordpress/element';
 
-import addBlockEditorDynamicStyles from '@Controls/addBlockEditorDynamicStyles';
 import scrollBlockToView from '@Controls/scrollBlockToView';
 import { useDeviceType } from '@Controls/getPreviewType';
 import { getFallbackNumber } from '@Controls/getAttributeFallback';
@@ -19,6 +18,8 @@ import responsiveConditionPreview from '@Controls/responsiveConditionPreview';
 import DynamicCSSLoader from '@Components/dynamic-css-loader';
 
 import { useSelect } from '@wordpress/data';
+import { compose } from '@wordpress/compose';
+import AddStaticStyles from '@Controls/AddStaticStyles';
 
 const PostTimelineComponent = ( props ) => {
 	const deviceType = useDeviceType();
@@ -80,7 +81,6 @@ const PostTimelineComponent = ( props ) => {
 
 	useEffect( () => {
 		// Replacement for componentDidUpdate.
-		addBlockEditorDynamicStyles();
 		const loadPostTimelineEditor = new CustomEvent( 'UAGTimelineEditor', {
 			// eslint-disable-line no-undef
 			detail: {},
@@ -180,4 +180,6 @@ const PostTimelineComponent = ( props ) => {
 		</>
 	);
 };
-export default PostTimelineComponent;
+export default compose(
+	AddStaticStyles,
+)( PostTimelineComponent );
